moving non function code into subdir
