+++ domain-search-by-owner/function.json
@@ -6,7 +6,7 @@
       "type": "httpTrigger",
       "direction": "in",
       "name": "req",
-      "methods": ["get", "post"]
+      "methods": ["post"]
     },
     {
       "type": "http",
