+++ package.json
@@ -13,14 +13,14 @@
   },
   "devDependencies": {
     "eslint": "8.15.0",
-    "eslint-plugin-prettier": "4.0.0",
-    "typescript": "4.0.0"
+    "eslint-plugin-prettier": "4.0.0"
   },
   "dependencies": {
     "@azure/functions": "3.0.0",
-    "@zero-tech/data-store": "0.0.2",
+    "@zero-tech/data-store-core": "^0.0.3",
     "ajv": "8.11.0",
     "env-var": "7.1.1",
-    "mongodb": "4.3.1"
+    "mongodb": "4.3.1",
+    "typescript": "^4.6.4"
   }
 }
