+++ src/helpers/index.ts
@@ -1,7 +1,7 @@
-import { getDbConnectionInfo } from "@zero-tech/data-store/lib/database/helpers";
+import { getDbConnectionInfo } from "@zero-tech/data-store-core/lib/database/helpers";
 import { Context } from "@azure/functions";
 import { MongoClient } from "mongodb";
-import { MongoDbService } from "@zero-tech/data-store/lib/database/mongo/mongoDbService";
+import { MongoDbService } from "@zero-tech/data-store-core/lib/database/mongo/mongoDbService";
 
 export async function getConnectedDbClient(context: Context) {
   const dbConnectionInfo = getDbConnectionInfo();
