+++ src/schemas.ts
@@ -1,4 +1,4 @@
-import { Address, DomainFindOptions, DomainId } from "@zero-tech/data-store";
+import { Address, DomainFindOptions, DomainId } from "@zero-tech/data-store-core";
 import Ajv, { JSONSchemaType } from "ajv";
 import { PartialSchema } from "ajv/dist/types/json-schema";
